requirements.txt
requests
beautifulsoup4
python-dotenv
psycopg[binary]
schedule
python-telegram-bot
playwright
flask